+++ docs/release_notes.rst
@@ -6,6 +6,19 @@ PUDL Release Notes
 v2024.X.X (2024-XX-XX)
 ---------------------------------------------------------------------------------------
 
+New Data Coverage
+^^^^^^^^^^^^^^^^^
+
+EIA AEO
+~~~~~~~
+
+* Added new tables from EIA AEO table 54:
+
+  * :ref:`core_eiaaeo__yearly_projected_fuel_cost_in_electric_sector_by_type`
+    contains generation capacity & generation projections for the electric
+    sector, broken out by technology type. See issue :issue:`3649` and PR
+    :pr:`3656`.
+
 .. _release-v2024.5.0:
 
 ---------------------------------------------------------------------------------------